+++ src/main/java/com/marcus/titan/infra/security/config/JWTUserData.java
@@ -4,5 +4,5 @@ package com.marcus.titan.infra.security.config;
 import lombok.Builder;
 
 @Builder
-public record JWTUserData(Long userId, String name) {
+public record JWTUserData(Integer userId, String name) {
 }
